ruff check and factorization
